src/ocifn-generate-jwt-assertion-python/requirements.txt
fdk>=0.1.39
PyJWT
cryptography